flask/requirements.txt
Click==8.1
Flask==2.2.2
Flask-MySQLdb==1.0.1
itsdangerous==2.1.0
Jinja2==3.0
MarkupSafe==2.1.2
uWSGI==2.0.21
Werkzeug==2.2.2
mysql-connector

